requirements.txt
jupyterlab==0.35.4
Keras==2.3.1
librosa==0.7.1
matplotlib==3.0.2
matplotlib2tikz==0.6.18
numpy==1.16.1
pandas==0.24.1
pydub==0.23.1
soundfile==0.10.3.post1
scikit-learn==0.20.2
scipy==1.2.1
SIDEKIT==1.3.2
tensorflow-gpu==2.0
tqdm==4.31.1
webrtcvad==2.0.10